Restructure loaders package dir; fix syntax error in packagejson
